project/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Werkzeug==3.0.1
google-cloud-logging==3.8.0
google-cloud-storage==2.10.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
feedparser==6.0.10
ebooklib==0.18
lxml==4.9.3
beautifulsoup4==4.12.2
Pillow==10.1.0
redis==5.0.1
Flask-Limiter==3.5.0